Create instance model outside xtext readonly in new resource set
Instantiation will try to resolve unresolved references. This leads to a
deadlock if the resource is already locked in the readonly action on the
xtext document. This should happen only if the AADL file containing the
root component implementation contains errors, e.g., uses properties
from a missing resource set.