WIP of Undo/Redo rework.
The ultimate goal is to replace the current undo capability that is is
very limited with an implementation that allow undoing and redoing all
actions if the text editor is not open. If a text editor is open then
capabilities will be degraded.

- Removed implementation of ICustomUndoRedoFeature. Graphiti undo/redo
no longer needed
- Removed ability to disable Graphiti modification listener. No longer
necessary with new undo implementation.
- Added undo support for add/remove element
- Added undo support for ghosting
- Added undo support for setDiagramConfiguration
- Added ActionService, AgeAction, and ActionExecutor
- Implemented global undo/redo stacks.
- Reworked interface between AgeDiagram, GraphitiAgeDiagram, and
AgeDiagramEditor to use AgeAction interface.
- Undo/Redo actions for an editor are invalidated if the editor is
closed.
- Implemented switching to appropriate editor when undo/redo occurs.