Restructured meta-model related to types, type sets, and type products.
The purpose of this restructuring was to create a specific EClass to
represent the results of flattening out a type set. This EClass is
TypeTokenInstance which is the super type of TypeInstance and
TypeProductInstance. TypeSetElement becomes the new super type of
TypeTokenInstance and TypeSetInstance.

Also created AbstractTypeSet and AnonymousTypeSet. Propagations now
point to an AnonymousTypeSet instead of a list of elements. This was
done to have a single implementation of flatten in AbstractTypeSet.