Lookup propagation instances by name instead of declarative reference.
This is necessary to find an overridden propagation. If an inherited error flow refers to an overridden propagation, the propagation instance cannot be found by looking at the declarative reference. This is because the ErrorFlow will point to the original ErrorPropagation while the ErrorPropagationInstance will point to the new ErrorPropagation.