Separated PropagationRefrence classes.
ConditionPropagationReference is used in a condition expression in which
a reference to a propagation can have a type set associated with it.
DestinationPropagationReference is used as the destination of an
outgoing propagation condition in which a reference to a propagation can
have a single type token associated with it, but not a type set.